backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.0
qdrant-client[fastembed]>=1.17.0
langchain-text-splitters>=0.3.0
pydantic>=2.0.0
